Move Http utilities into separate file
